// src/api/generated/types.gen.ts
export type ModelsChallengeDetailsResponse = {
    author?: string;
    category?: string;
    createdAt?: string;
    description?: string;
    difficulty?: string;
    file_url?: string;
    id?: string;
    name?: string;
    points?: number;
    solves?: number;
    source?: string;
    stage?: string;
};